refactor: 이미지 최적화 - png에서 svg 변경.
